Remove setting LAT/LONG in the container
Given that all of FA's official binaries and scripts do not set the
LAT/LONG on the app side (they never even ask for a user's location
doing this is clearly unnecessary - and for many people it seems to
prevenet MLAT from working correctly.
Simply set your location on the FA website and all will be fine.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>